requirements.txt
langchain==0.1.0
openai==1.8.0
faiss-cpu==1.7.4
pydantic==2.5.3
python-dotenv==1.0.0
pypdf2==3.0.1
pdfminer.six==20221105
email-validator==2.1.0.post1
numpy==1.24.3
pandas==2.1.4
fastapi==0.109.0
uvicorn==0.25.0
python-multipart==0.0.6 